requirements.txt
Django==5.2.4
django-unfold==0.63.0
pillow==11.3.0
python-decouple==3.8
django-import-export==4.3.9
openpyxl==3.1.5
xlsxwriter==3.2.0
pandas==2.3.1 